High availability not highly available
The term highly available often implies the deployment of infrastructure
with redundant components. Many publication servers (I think the NCC one
is the exception) do not support the deployment of redundant/highly
available instances.

I tried to rephrase it to (a) clarify the operational goal and (b) avoid
the HA term